backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
pandas==1.5.3
openpyxl==3.1.2
python-dotenv==0.19.0
easyocr==1.7.1
pillow==10.1.0
numpy==1.26.2
opencv-python-headless==4.8.1.78
aiohttp==3.9.1
sqlalchemy==1.4.23
pyjwt==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 